requirements.txt
pytest
pylint
requests
psycopg2-binary
python-dotenv
streamlit
pandas
numpy
plotly